Refactor file structure and rename components for codex-chat feature
- Move chat components to pages/codex-chat directory for better organization
- Rename App component to CodexChatPage for clarity
- Update imports and paths to reflect new structure
- Adjust layout dimensions in CodexChatPage
- Clean up type definitions and imports

This reorganization improves the project structure by grouping related chat components together under a clear feature-based hierarchy.
