+++ webview-ui/src/index.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client';
-import { App } from './features/app';
+import { CodexChatPage } from './pages/codex-chat/page';
 
 const container = document.getElementById('root')!;
 const root = createRoot(container);
-root.render(<App />);
+root.render(<CodexChatPage />);
